test: 공통 renderHook 및 wrapper 분리
