+++ src/__tests__/useCart.test.tsx
@@ -1,11 +1,8 @@
-import { act, renderHook, waitFor } from "@testing-library/react";
-import { MemoryRouter } from "react-router-dom";
+import { act, waitFor } from "@testing-library/react";
 import { deleteCartItem } from "../apis/cartItems/deleteCartItem";
 import { getCartItems } from "../apis/cartItems/getCartItems";
 import { patchCartItem } from "../apis/cartItems/patchCartItem";
-import { CartProvider } from "../contexts/CartContext";
-import useCart from "../hooks/useCart";
-import { ToastProvider } from "../contexts/ToastContext";
+import { renderCartHook } from "../test-utils/renderCartHook";
 
 jest.mock("../apis/httpClient", () => ({
   API_KEY: "mock-api-key",
@@ -41,14 +38,6 @@ const mockCartItems = [
   },
 ];
 
-const wrapper = ({ children }: React.PropsWithChildren) => (
-  <MemoryRouter>
-    <ToastProvider>
-      <CartProvider>{children}</CartProvider>
-    </ToastProvider>
-  </MemoryRouter>
-);
-
 describe("useCart 훅 테스트", () => {
   beforeEach(() => {
     jest.clearAllMocks();
@@ -57,7 +46,7 @@ describe("useCart 훅 테스트", () => {
 
   describe("초기 렌더링했을 때", () => {
     it("전체 장바구니 상품 데이터가 불러와지고, 원본 상태와 파생 상태가 초기화된다.", async () => {
-      const { result } = renderHook(() => useCart(), { wrapper });
+      const { result } = renderCartHook();
 
       await waitFor(() => {
         expect(getCartItems).toHaveBeenCalled();
@@ -70,10 +59,10 @@ describe("useCart 훅 테스트", () => {
     });
 
     it("전체 상품이 선택된다.", async () => {
-      const { result } = renderHook(() => useCart(), { wrapper });
+      const { result } = renderCartHook();
 
       await waitFor(() => {
-        expect(result.current.allChecked).toBe(true);
+        expect(result.current.isAllChecked).toBe(true);
         result.current.cartItemsCheckData.forEach((item) => {
           expect(item.checked).toBe(true);
         });
@@ -84,7 +73,7 @@ describe("useCart 훅 테스트", () => {
 
   describe("특정 상품의 삭제 버튼을 클릭했을 때", () => {
     it("특정 상품이 장바구니에서 삭제된다.", async () => {
-      const { result } = renderHook(() => useCart(), { wrapper });
+      const { result } = renderCartHook();
 
       await waitFor(() => {
         expect(result.current.cartItemsData).toHaveLength(mockCartItems.length);
@@ -99,7 +88,8 @@ describe("useCart 훅 테스트", () => {
     });
 
     it("삭제한 상품의 가격만큼 주문 금액 정보가 변경된다.", async () => {
-      const firstItemPrice = mockCartItems[0].product.price;
+      const firstItemPrice =
+        mockCartItems[0].product.price * mockCartItems[0].quantity;
       const secondItemPrice =
         mockCartItems[1].product.price * mockCartItems[1].quantity;
       const totalPrice = firstItemPrice + secondItemPrice;
@@ -108,7 +98,7 @@ describe("useCart 훅 테스트", () => {
         .mockResolvedValueOnce(mockCartItems)
         .mockResolvedValueOnce([mockCartItems[1]]);
 
-      const { result } = renderHook(() => useCart(), { wrapper });
+      const { result } = renderCartHook();
 
       await waitFor(() => {
         expect(result.current.orderPrice).toBe(totalPrice);
@@ -121,7 +111,6 @@ describe("useCart 훅 테스트", () => {
       expect(result.current.orderPrice).toBe(secondItemPrice);
     });
   });
-
   describe("특정 상품의 개수를 증가시켰을 때", () => {
     it("특정 상품의 수량이 증가한다.", async () => {
       const initialQuantity = mockCartItems[0].quantity;
@@ -136,7 +125,7 @@ describe("useCart 훅 테스트", () => {
         .mockResolvedValueOnce(mockCartItems)
         .mockResolvedValueOnce(updatedCartItems);
 
-      const { result } = renderHook(() => useCart(), { wrapper });
+      const { result } = renderCartHook();
 
       await waitFor(() => {
         expect(result.current.cartItemsData[0].quantity).toBe(initialQuantity);
@@ -155,209 +144,207 @@ describe("useCart 훅 테스트", () => {
         initialQuantity + 1
       );
     });
-
-    it("수량을 증가시킨 상품의 가격만큼 주문 금액 정보가 변경된다.", async () => {
-      const initialQuantity = mockCartItems[0].quantity;
-      const productPrice = mockCartItems[0].product.price;
-
-      const updatedCartItems = [...mockCartItems];
-      updatedCartItems[0] = {
-        ...updatedCartItems[0],
-        quantity: initialQuantity + 1,
-      };
-
-      (getCartItems as jest.Mock)
-        .mockResolvedValueOnce(mockCartItems)
-        .mockResolvedValueOnce(updatedCartItems);
-
-      const { result } = renderHook(() => useCart(), { wrapper });
-
-      const initialTotal = mockCartItems.reduce(
-        (sum, item) => sum + item.quantity * item.product.price,
-        0
-      );
-
-      await waitFor(() => {
-        expect(result.current.orderPrice).toBe(initialTotal);
-      });
-
-      await act(async () => {
-        await result.current.increaseItemQuantity(101, initialQuantity);
-      });
-
-      expect(result.current.orderPrice).toBe(initialTotal + productPrice);
-    });
   });
 
-  describe("특정 상품의 개수를 감소시켰을 때", () => {
-    it("특정 상품의 수량이 감소한다.", async () => {
-      const initialQuantity = mockCartItems[1].quantity;
-      expect(initialQuantity).toBeGreaterThan(1);
+  it("수량을 증가시킨 상품의 가격만큼 주문 금액 정보가 변경된다.", async () => {
+    const initialQuantity = mockCartItems[0].quantity;
+    const productPrice = mockCartItems[0].product.price;
 
-      const updatedCartItems = [...mockCartItems];
-      updatedCartItems[1] = {
-        ...updatedCartItems[1],
-        quantity: initialQuantity - 1,
-      };
+    const updatedCartItems = [...mockCartItems];
+    updatedCartItems[0] = {
+      ...updatedCartItems[0],
+      quantity: initialQuantity + 1,
+    };
 
-      (getCartItems as jest.Mock)
-        .mockResolvedValueOnce(mockCartItems)
-        .mockResolvedValueOnce(updatedCartItems);
+    (getCartItems as jest.Mock)
+      .mockResolvedValueOnce(mockCartItems)
+      .mockResolvedValueOnce(updatedCartItems);
 
-      const { result } = renderHook(() => useCart(), { wrapper });
+    const { result } = renderCartHook();
 
-      await waitFor(() => {
-        expect(result.current.cartItemsData[1].quantity).toBe(initialQuantity);
-      });
+    const initialTotal = mockCartItems.reduce(
+      (sum, item) => sum + item.quantity * item.product.price,
+      0
+    );
 
-      await act(async () => {
-        await result.current.decreaseItemQuantity(102, initialQuantity);
-      });
-
-      expect(patchCartItem).toHaveBeenCalledWith({
-        cartId: 102,
-        quantity: initialQuantity - 1,
-      });
-
-      expect(result.current.cartItemsData[1].quantity).toBe(
-        initialQuantity - 1
-      );
+    await waitFor(() => {
+      expect(result.current.orderPrice).toBe(initialTotal);
     });
 
-    it("수량을 감소시킨 상품의 가격만큼 주문 금액 정보가 변경된다.", async () => {
-      const initialQuantity = mockCartItems[1].quantity;
-      const productPrice = mockCartItems[1].product.price;
-
-      const updatedCartItems = [...mockCartItems];
-      updatedCartItems[1] = {
-        ...updatedCartItems[1],
-        quantity: initialQuantity - 1,
-      };
-
-      (getCartItems as jest.Mock)
-        .mockResolvedValueOnce(mockCartItems)
-        .mockResolvedValueOnce(updatedCartItems);
-
-      const { result } = renderHook(() => useCart(), { wrapper });
-
-      const initialTotal = mockCartItems.reduce(
-        (sum, item) => sum + item.quantity * item.product.price,
-        0
-      );
-
-      await waitFor(() => {
-        expect(result.current.orderPrice).toBe(initialTotal);
-      });
-
-      await act(async () => {
-        await result.current.decreaseItemQuantity(102, initialQuantity);
-      });
-
-      expect(result.current.orderPrice).toBe(initialTotal - productPrice);
+    await act(async () => {
+      await result.current.increaseItemQuantity(101, initialQuantity);
     });
+
+    expect(result.current.orderPrice).toBe(initialTotal + productPrice);
+  });
+});
+
+describe("특정 상품의 개수를 감소시켰을 때", () => {
+  it("특정 상품의 수량이 감소한다.", async () => {
+    const initialQuantity = mockCartItems[1].quantity;
+    expect(initialQuantity).toBeGreaterThan(1);
+
+    const updatedCartItems = [...mockCartItems];
+    updatedCartItems[1] = {
+      ...updatedCartItems[1],
+      quantity: initialQuantity - 1,
+    };
+
+    (getCartItems as jest.Mock)
+      .mockResolvedValueOnce(mockCartItems)
+      .mockResolvedValueOnce(updatedCartItems);
+
+    const { result } = renderCartHook();
+
+    await waitFor(() => {
+      expect(result.current.cartItemsData[1].quantity).toBe(initialQuantity);
+    });
+
+    await act(async () => {
+      await result.current.decreaseItemQuantity(102, initialQuantity);
+    });
+
+    expect(patchCartItem).toHaveBeenCalledWith({
+      cartId: 102,
+      quantity: initialQuantity - 1,
+    });
+
+    expect(result.current.cartItemsData[1].quantity).toBe(initialQuantity - 1);
   });
 
-  describe("전체 선택 버튼 클릭했을 때", () => {
-    it("전체 선택이 되어있는 경우 전체 선택을 취소하면, 모든 상품의 개별 선택 여부가 false로 변경된다.", async () => {
-      const { result } = renderHook(() => useCart(), { wrapper });
+  it("수량을 감소시킨 상품의 가격만큼 주문 금액 정보가 변경된다.", async () => {
+    const initialQuantity = mockCartItems[1].quantity;
+    const productPrice = mockCartItems[1].product.price;
 
-      await waitFor(() => {
-        expect(result.current.allChecked).toBe(true);
-      });
+    const updatedCartItems = [...mockCartItems];
+    updatedCartItems[1] = {
+      ...updatedCartItems[1],
+      quantity: initialQuantity - 1,
+    };
 
-      act(() => {
-        result.current.toggleAllChecked();
-      });
+    (getCartItems as jest.Mock)
+      .mockResolvedValueOnce(mockCartItems)
+      .mockResolvedValueOnce(updatedCartItems);
 
-      expect(result.current.allChecked).toBe(false);
-      result.current.cartItemsCheckData.forEach((item) => {
-        expect(item.checked).toBe(false);
-      });
+    const { result } = renderCartHook();
 
-      expect(result.current.orderPrice).toBe(0);
-      expect(result.current.shippingFee).toBe(0);
-      expect(result.current.totalPrice).toBe(0);
+    const initialTotal = mockCartItems.reduce(
+      (sum, item) => sum + item.quantity * item.product.price,
+      0
+    );
+
+    await waitFor(() => {
+      expect(result.current.orderPrice).toBe(initialTotal);
     });
 
-    it("전체 선택이 되어 있지 않은 경우 전체 선택을 하면, 모든 상품의 개별 선택 여부가 true로 변경된다.", async () => {
-      const { result } = renderHook(() => useCart(), { wrapper });
-
-      await waitFor(() => {
-        expect(result.current.cartItemsData).toHaveLength(mockCartItems.length);
-      });
-
-      act(() => {
-        result.current.toggleAllChecked();
-      });
-
-      expect(result.current.allChecked).toBe(false);
-
-      act(() => {
-        result.current.toggleAllChecked();
-      });
-
-      expect(result.current.allChecked).toBe(true);
-      result.current.cartItemsCheckData.forEach((item) => {
-        expect(item.checked).toBe(true);
-      });
-
-      const expectedTotal = mockCartItems.reduce(
-        (sum, item) => sum + item.quantity * item.product.price,
-        0
-      );
-
-      expect(result.current.orderPrice).toBe(expectedTotal);
+    await act(async () => {
+      await result.current.decreaseItemQuantity(102, initialQuantity);
     });
+
+    expect(result.current.orderPrice).toBe(initialTotal - productPrice);
+  });
+});
+
+describe("전체 선택 버튼 클릭했을 때", () => {
+  it("전체 선택이 되어있는 경우 전체 선택을 취소하면, 모든 상품의 개별 선택 여부가 false로 변경된다.", async () => {
+    const { result } = renderCartHook();
+
+    await waitFor(() => {
+      expect(result.current.isAllChecked).toBe(true);
+    });
+
+    act(() => {
+      result.current.toggleAllChecked();
+    });
+
+    expect(result.current.isAllChecked).toBe(false);
+    result.current.cartItemsCheckData.forEach((item) => {
+      expect(item.checked).toBe(false);
+    });
+
+    expect(result.current.orderPrice).toBe(0);
+    expect(result.current.shippingFee).toBe(0);
+    expect(result.current.totalPrice).toBe(0);
   });
 
-  describe("특정 상품의 선택 버튼을 클릭했을 때", () => {
-    it("이미 선택이 되어있는 경우 취소하면, 선택한 상품의 가격만큼 주문 금액 정보가 변경된다.", async () => {
-      const { result } = renderHook(() => useCart(), { wrapper });
+  it("전체 선택이 되어 있지 않은 경우 전체 선택을 하면, 모든 상품의 개별 선택 여부가 true로 변경된다.", async () => {
+    const { result } = renderCartHook();
 
-      await waitFor(() => {
-        expect(result.current.cartItemsData).toHaveLength(mockCartItems.length);
-      });
-
-      const firstItemPrice =
-        mockCartItems[0].quantity * mockCartItems[0].product.price;
-      const initialTotal = mockCartItems.reduce(
-        (sum, item) => sum + item.quantity * item.product.price,
-        0
-      );
-
-      act(() => {
-        result.current.toggleItemChecked(101);
-      });
-
-      expect(result.current.getItemChecked(101)).toBe(false);
-      expect(result.current.getItemChecked(102)).toBe(true);
-
-      expect(result.current.orderPrice).toBe(initialTotal - firstItemPrice);
+    await waitFor(() => {
+      expect(result.current.cartItemsData).toHaveLength(mockCartItems.length);
     });
 
-    it("선택이 되어있지 않은 경우 선택하면, 선택한 상품의 가격만큼 주문 금액 정보가 변경된다.", async () => {
-      const { result } = renderHook(() => useCart(), { wrapper });
-
-      await waitFor(() => {
-        expect(result.current.cartItemsData).toHaveLength(mockCartItems.length);
-      });
-
-      act(() => {
-        result.current.toggleAllChecked();
-      });
-
-      expect(result.current.orderPrice).toBe(0);
-
-      act(() => {
-        result.current.toggleItemChecked(101);
-      });
-
-      expect(result.current.getItemChecked(101)).toBe(true);
-      expect(result.current.getItemChecked(102)).toBe(false);
-
-      const firstItemPrice =
-        mockCartItems[0].quantity * mockCartItems[0].product.price;
-      expect(result.current.orderPrice).toBe(firstItemPrice);
+    act(() => {
+      result.current.toggleAllChecked();
     });
+
+    expect(result.current.isAllChecked).toBe(false);
+
+    act(() => {
+      result.current.toggleAllChecked();
+    });
+
+    expect(result.current.isAllChecked).toBe(true);
+    result.current.cartItemsCheckData.forEach((item) => {
+      expect(item.checked).toBe(true);
+    });
+
+    const expectedTotal = mockCartItems.reduce(
+      (sum, item) => sum + item.quantity * item.product.price,
+      0
+    );
+
+    expect(result.current.orderPrice).toBe(expectedTotal);
+  });
+});
+
+describe("특정 상품의 선택 버튼을 클릭했을 때", () => {
+  it("이미 선택이 되어있는 경우 취소하면, 선택한 상품의 가격만큼 주문 금액 정보가 변경된다.", async () => {
+    const { result } = renderCartHook();
+
+    await waitFor(() => {
+      expect(result.current.cartItemsData).toHaveLength(mockCartItems.length);
+    });
+
+    const firstItemPrice =
+      mockCartItems[0].quantity * mockCartItems[0].product.price;
+    const initialTotal = mockCartItems.reduce(
+      (sum, item) => sum + item.quantity * item.product.price,
+      0
+    );
+
+    act(() => {
+      result.current.toggleItemChecked(101);
+    });
+
+    expect(result.current.getItemChecked(101)).toBe(false);
+    expect(result.current.getItemChecked(102)).toBe(true);
+
+    expect(result.current.orderPrice).toBe(initialTotal - firstItemPrice);
+  });
+
+  it("선택이 되어있지 않은 경우 선택하면, 선택한 상품의 가격만큼 주문 금액 정보가 변경된다.", async () => {
+    const { result } = renderCartHook();
+
+    await waitFor(() => {
+      expect(result.current.cartItemsData).toHaveLength(mockCartItems.length);
+    });
+
+    act(() => {
+      result.current.toggleAllChecked();
+    });
+
+    expect(result.current.orderPrice).toBe(0);
+
+    act(() => {
+      result.current.toggleItemChecked(101);
+    });
+
+    expect(result.current.getItemChecked(101)).toBe(true);
+    expect(result.current.getItemChecked(102)).toBe(false);
+
+    const firstItemPrice =
+      mockCartItems[0].quantity * mockCartItems[0].product.price;
+    expect(result.current.orderPrice).toBe(firstItemPrice);
   });
 });
